+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Fix `__has_include` check in ASLog.h [Philipp Smorygo](Philipp.Smorygo@jetbrains.com)
 - Fix potential deadlock in ASControlNode [Garrett Moon](https://github.com/garrettmoon)
 - [Yoga Beta] Improvements to the experimental support for Yoga layout [Scott Goodson](appleguy)
+- Fix inserting a sublayer in an index of NSNotFound (Michael Schneider)[https://github.com/maicki] (#80)(https://github.com/TextureGroup/Texture/pull/80)